Hacky balances at-data implemented.
Currently, the balances are selected for the current date, so when we
try to get a balance for a time where the account didn't exist, we
get an error. This is bad behaviour and we should make sure that we
only select accounts that exist at that time.

It seems like it now would be a good time to implement some kind of
query filtering when selecting accounts.