Add `account open` command.
This command opens an account and inserts a balance at the opening date.

There are issues with adding the same flag to different subommands, so
I had to move the currency flag to be on the account command as a
persistent flag. This doesn't seem right and it could just be an issue
with the way that viper and cobra are being used together.