drone-functionaltests: improve testing feedback
Previously, there was an uneccesary log that was happening everytime
the functional test were run. The log didn't actually give any
information so it has been removed.

Some more useful feedback has been provided when the store cannot be
contacted when setting up there tests, too. Previously it would say
that it couldn't connect to the store but not give any information.
Now, it tells you the address that it was trying to connect to, which
helped when the test container name was changed but the address for the
client wasn't updated.