serverrouterrefactor: Refactor to use unassigned ports
Prior to this change, each test would attempt to use a different port to
server on, starting from 23456. Although the chances of this causing
issues were really slim, it wasn't following best practises.

This change provides a factory for creating the components that are used
within the test, each configured to serve, listen or call on the
unassigned port.