rename some resources with mon at the start
Prior to this change, in the host environment, it wasn't explicitly clear
that some of the timages that had been made were for this project. For
example, when a specific postgres image was created, it wasn't easy to
tell that it was from the mon project when it only had the name postgres

This change prepends mon to the start of some resources to make this a
bit clearer.