Don't have version of image inside Makefile
Before this commit, the version to test against was coded into the
makefile. This caused some issues when testing sometimes as an accidental
configuration would mean that the wrong version woudl be tested. It's
best to inject the version through some automated system or have the
user put the version in when they are doing it manually.