Increase threshold for date flag
Prior to this change, when running the test on a low-power machine, they
would often fail because the time threshold, for when a flag parses a
time as yesterday, was too small.

This would mean that the test wouldn't complete quick enough to go from
test invocation to flag parsing.

This change increases the threshold to 10 milliseconds from 5.