accountfilters: Improve comments and restructure file
Prior to this commit, the comments for a couple of the filter functions
were inconsisten in style compared to each other.

This commit ensures that the functions are in the correct order, making
similar or related functions next to each other and also improves the
comments so that they match the style of each other.