[sortbyid] Create table.AccountBalance type
Priot to this commit, when printing a table of accounts with balances,
the values would be passed in as a map. This would mean that the order
in which the accounts would be printed would be randomised by go.

This commit creates an AccountBalance type. A slice of AccountBalance
is now received as a parameter to the table printing function, meaning
that the order in which they are in the slice that is passed is
preserved when printing them out.