Sort by balance amount
Prior to this commit, there was an AccountBalance type within the table
package. Now that this type is required in more than 1 package, it
has been extracted to another place within the project. This should not
be the final place, though, as there will be a better place for it after
some thought and experience of using it, I hope.

This commit also adds a function which will sort a slice of
AccountBalance by the value of the Balance.Amount