Add flag for sort
Priot to this commit, the validation for the value that the sort-by flag
gets set to was done within the cmb package, which made the command more
complex by providing it with too much responsibility.

This commit creates a sort.Key type which can be used as a pflag.Value,
moving the validation of the value to within the sort package, rather
than in the accounts command.