Add type format with a parse method
Prior to this commit, parsing an explicit date used a package constant.

This commit adds a type called format with a parse method which can be
used in the slice of parsers. This way, we can easily reuse the format
type with different formats and add them to the slice very easily.