Split date.Set tests into logical groups
Prior to this commit, the tests for setting the flag value were all
within a single large test table.

This commit splits up the different types of tests – for errors,
explicit and relative dates – into into their own test functions.

This means that the body of the test functions has become a lot simpler.