Add reopen account subcommand
Prior to this commit, the user would have to use the moncli account
update command and manually keep some of the values as they are already
set.

This commit adds a command which takes the id of an account as an
argument and keeps all values except for the closed date the same.