Add dirty to version name if there is WIP
Prior to this commit, if a build was made whilst there was work in
progress in the repository, the version would not show this at all.

This change adds the '--dirty' flag to the git command which is used
to create the version name. This flag means that the phrase 'dirty'
is added if there is work in progress.