Set golang build image tag to 1.10.4
Prior to this commit, the latest version of the golang image was
causing the unit tests to fail, perhaps because there were directories
being tested that didn't have any tests in.

This change sets the version to 1.10.4, which is the version of go that
was used to develop the changes.